Implement dual decomposition
* Implement basic run optimization step.

* Fix std::move and rvalue passing through ctor bug. Bug was discovered
using valgrind.

* Introduce dual decomposition constraints into the
PrimalDualMinCutSolver. This is along with all necesary update rules.

* Introduce proof-of-concept reuse of trees during maxflow along with
proper marking of nodes to make the reuse of trees possible.

* Add ctors in structs to enable `emplace_back` to work as expected.

* Introduce the concept of a `stable_vector` to prevent the
"reallocation" bug that happens with std::vector and complex types that
can not be easily reallocated and recreated.

* Introduce a `runOptimizationStep` in the dual decomposition to enable
a rudimentary form of dual decomposition subgradient ascent.

* Create a driver program that would implement rudimentary dual
decomposition subgradient ascent on DIMACS format maxflow graphs.