Implement different ideas to improve convergence
* Use 64 bit precision for all capacities to push scaling allowed.
Result shows that this only grants a few more levels of scale and does
not guarantee convergence.

* Implement a momentum term in subgradient ascent. This significantly
improves the convergence rate, but makes detecting cycles trickier.

* Implement and experiment with tracking different statistics within
each scale's iteration. Settle on detecting if the maximum lower bound
increases between two independent groups of iterations (to detect a slow
down in convergence and terminate scale.)

* Run a test to see if decoding the mincut using a manual reverse BFS
from sink after maxflow would yield a more stable convergence by
"guaranteeing" a minimal cardinality mincut (ie. a mincut with the most
source nodes possible.) Results seem to indicate this is identical to
the results returned by the typical way to decode a mincut.

* Remove previous regularization performed on each node in favor of the
regularization discussed in the point directly above.

* A general note: this is considered a work-in-progress commit.
Refactoring is expected in the future to clean up R&D code introduced in
this commit.