Implement additional convergence and decoding techniques
* Implement narrow-band decoding. This greatly improves the rate at
which a primal feasible solution can be decoded. The narrow-band
decoding focuses a min-cut decoding centered on the nodes that currently
disagree in the dual decomposition. The memory usage of the initial
breadth-first search used to determine the narrow-band used needs to be
addressed in the future.

* Implement full graph constraints on constrained nodes in the dual
decomposition. Hypothetically, this should constrain the problem further
and possibly give better convergence. In testing and practice, there
seems to be little difference between this and the linear chain graph
used beforehand.