Implement a scalable version of graph functions
* Implement a CSR representation of a graph using memory maps so that
large or huge graphs can be stored with a limited amount of memory.

* Use CSR representation to implement METIS graph partitioning.

* Use CSR representation to implement the narrow-band decoding
heuristic.

* Add boost dependecy to CMakeLists (at version used in development.)
Require the threads dependency.

* Remove the use of `stable_vector` from the dual decomposition in favor
of a vector of std::unique_ptr.

* Track the greatest lower bound (as a double normalized by the scale)
computed in the dual decomposition as well as the current nodes in
disagreement as part of the computation.

* Implement a version of `solve` in the dual decomposition that allows a
custom decoder to be passed in (as plumbing to enable the narrow-band
heuristic.)

* Reenable the use of the existing narrow-band heuristic for comparison
(this will be removed shortly in the future.)

* Refactor the way that dimacs graphs were read in so that it can be
adapted more easily in the future.

* Implement a general MmapArray class that stores data as a
memory-mapped pointer.