Introduce regularization on constrainied nodes
* Introduce regularization on constrainied nodes to ensure convergence.

* Theory states that as long as the scale of the minimum cut
(sub)problem is great enough, introducing a regularization penalty on
constrained nodes which do not converge to S (zero) in the min cut will
not affect the lower bound (since the min cut in the regularized problem
will be guaranteed to be a min cut in the original problem.) This is
turn allows the dual decomposition to converge.

* Introduce a warning message when the threshold afterwich the lower
bound guarantee is no longer valid.